Release 3.0.0-beta.38

+++ bin/github-tag-clean.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
+TOP_DIR=$(pwd)
+VERSION=$(node -p -e "require('$TOP_DIR/package.json').version")
 for tagName in $(git tag | grep 'alpha\|beta')
 do
-  git push --delete origin $tagName
-  git tag -d $tagName
+  if [[ "$tagName" != "$VERSION" ]]
+  then
+    git push --delete origin $tagName
+    git tag -d $tagName
+  fi
 done
